requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.8
django-bootstrap4==3.0.1
djangorestframework==3.12.4
numpy==1.21.3
psycopg2==2.9.1
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
